move toolbarButton component to extra directory
