+++ packages/grafana-ui/src/components/Button/index.ts
@@ -1,4 +1,2 @@
 export * from './Button';
 export { ButtonGroup } from './ButtonGroup';
-export { ToolbarButton, ToolbarButtonVariant } from './ToolbarButton';
-export { ToolbarButtonRow } from './ToolbarButtonRow';
