+++ packages/grafana-ui/src/components/ToolbarButton/index.ts
@@ -1,0 +1,2 @@
+export * from './ToolbarButton';
+export { ToolbarButtonRow } from './ToolbarButtonRow';
